+++ src/map/next_js/interactive_helpers/commandHandler.ts
@@ -3,8 +3,17 @@ import path from "path";
 import commandHelpers from "./commandHelpers.js";
 import { helpMenu } from "./helpMenu.js";
 import { printFunction } from "./printer.js";
+import { State } from "../interactive.js";
+import { Widgets } from "blessed";
 
-function handleCommand(text, state, ui) {
+interface Screen {
+    screen: any;
+    titleBox?: Widgets.BoxElement;
+    outputBox: Widgets.Log;
+    inputBox: Widgets.TextboxElement;
+}
+
+async function handleCommand(text: string, state: State, ui: Screen) {
     const { chunks } = state;
     const { outputBox, inputBox, screen } = ui;
 
@@ -57,7 +66,7 @@ function handleCommand(text, state, ui) {
                 );
                 state.lastCommandStatus = true;
             } else {
-                outputBox.log(chalk.red(option), "is not a valid option");
+                outputBox.log(chalk.red(option) + " is not a valid option");
                 state.lastCommandStatus = false;
             }
         }
@@ -75,9 +84,10 @@ function handleCommand(text, state, ui) {
                 const funcId = text.split(" ")[2];
                 // check if the function exists
                 if (chunks[funcId]) {
-                    const funcCode = commandHelpers.getFunctionCode(
+                    const funcCode = await commandHelpers.getFunctionCode(
                         chunks,
-                        funcId
+                        funcId,
+                        state
                     );
                     printFunction(
                         outputBox,
@@ -102,17 +112,27 @@ function handleCommand(text, state, ui) {
                             state.functionNavHistory[
                                 state.functionNavHistoryIndex
                             ];
-                        const funcCode = commandHelpers.getFunctionCode(
-                            chunks,
-                            funcId
-                        );
-                        printFunction(
-                            outputBox,
-                            funcCode,
-                            chunks[funcId].description,
-                            state.funcWriteFile
-                        );
-                        state.lastCommandStatus = true;
+
+                        if (Object.keys(chunks).includes(funcId)) {
+                            const funcCode =
+                                await commandHelpers.getFunctionCode(
+                                    chunks,
+                                    funcId,
+                                    state
+                                );
+                            printFunction(
+                                outputBox,
+                                funcCode,
+                                chunks[funcId].description,
+                                state.funcWriteFile
+                            );
+                            state.lastCommandStatus = true;
+                        } else {
+                            outputBox.log(
+                                chalk.red(`No function with ID ${funcId} found`)
+                            );
+                            state.lastCommandStatus = false;
+                        }
                     } else {
                         outputBox.log(chalk.red("No previous function found"));
                         state.lastCommandStatus = false;
@@ -132,17 +152,26 @@ function handleCommand(text, state, ui) {
                             state.functionNavHistory[
                                 state.functionNavHistoryIndex
                             ];
-                        const funcCode = commandHelpers.getFunctionCode(
-                            chunks,
-                            funcId
-                        );
-                        printFunction(
-                            outputBox,
-                            funcCode,
-                            chunks[funcId].description,
-                            state.funcWriteFile
-                        );
-                        state.lastCommandStatus = true;
+                        if (Object.keys(chunks).includes(funcId)) {
+                            const funcCode =
+                                await commandHelpers.getFunctionCode(
+                                    chunks,
+                                    funcId,
+                                    state
+                                );
+                            printFunction(
+                                outputBox,
+                                funcCode,
+                                chunks[funcId].description,
+                                state.funcWriteFile
+                            );
+                            state.lastCommandStatus = true;
+                        } else {
+                            outputBox.log(
+                                chalk.red(`No function with ID ${funcId} found`)
+                            );
+                            state.lastCommandStatus = false;
+                        }
                     } else {
                         outputBox.log(chalk.red("No next function found"));
                         state.lastCommandStatus = false;
@@ -152,7 +181,7 @@ function handleCommand(text, state, ui) {
                     state.lastCommandStatus = false;
                 }
             } else {
-                outputBox.log(chalk.red(funcName), "is not a valid option");
+                outputBox.log(chalk.red(funcName) + " is not a valid option");
                 state.lastCommandStatus = false;
             }
         }
@@ -174,8 +203,23 @@ function handleCommand(text, state, ui) {
                     )
                 );
                 state.lastCommandStatus = true;
+            } else if (option === "writeimports") {
+                // modify the var in state
+                const modifyVal = text.split(" ")[2];
+                if (modifyVal === "true") {
+                    state.writeimports = true;
+                    outputBox.log("writeimports: " + chalk.green("true"));
+                    state.lastCommandStatus = true;
+                } else if (modifyVal === "false") {
+                    state.writeimports = false;
+                    outputBox.log("writeimports: " + chalk.yellow("false"));
+                    state.lastCommandStatus = true;
+                } else {
+                    outputBox.log(chalk.magenta(helpMenu.set));
+                    state.lastCommandStatus = false;
+                }
             } else {
-                outputBox.log(chalk.red(option), "is not a valid option");
+                outputBox.log(chalk.red(option) + " is not a valid option");
                 state.lastCommandStatus = false;
             }
         }
@@ -196,7 +240,7 @@ function handleCommand(text, state, ui) {
             }
         }
     } else {
-        outputBox.log(chalk.red(text), "is not a valid command");
+        outputBox.log((chalk.red(text), "is not a valid command"));
         state.lastCommandStatus = false;
     }
     inputBox.clearValue();