+++ src/strings/index.ts
@@ -138,6 +138,7 @@ const strings = async (
                         ast = parser.parse(jsCode, {
                             sourceType: "unambiguous",
                             plugins: ["jsx", "typescript"],
+                            errorRecovery: true,
                         });
                     } catch (err) {
                         continue;
@@ -155,6 +156,7 @@ const strings = async (
             const ast = parser.parse(fileContent, {
                 sourceType: "unambiguous",
                 plugins: ["jsx", "typescript"],
+                errorRecovery: true,
             });
 
             all_strings[file] = extractStrings(ast);